made random thinking algorithm

+++ AAI-Final-Assignment-WinForms/Entities/ItemSpawn.cs
@@ -10,7 +10,7 @@ namespace AAI_Final_Assignment_WinForms.Entities
 {
     public class ItemSpawn : BaseGameEntity
     {
-        public ItemSpawn(Vector2D pos, GameWorld world, float scale, int textureWidth, int textureHeight, double radius) :
+        public ItemSpawn(Vector2D pos, GameWorld world, float scale, int textureWidth, int textureHeight, float radius) :
             base(pos, world, scale, textureWidth, textureHeight, radius) {
 
         }
